third_party/WebKit/LayoutTests/fast/css/getComputedStyle/computed-style-grid-layout.html

<!DOCTYPE html>
<html>
<script>
var propertiesToTest = {
  "grid-auto-columns": true,
  "grid-auto-flow": true,
  "grid-auto-rows": true,
  "grid-area": true,
  "grid-column": true,
  "grid-column-end": true,
  "grid-column-start": true,
  "grid-template-columns": true,
  "grid-template-rows": true,
  "grid-row": true,
  "grid-row-end": true,
  "grid-row-start": true,
  "grid-template-areas": true,
};

function test()
{
    if (window.testRunner)
        testRunner.dumpAsText();
    var style = document.defaultView.getComputedStyle(document.body, "");
    var text = "";
    for (var i = 0; i != style.length; ++i) {
        var name = style.item(i);
        if (!propertiesToTest[name])
           continue;
        text += name + ": " + style.getPropertyValue(name) + ";\n";
    }
    document.getElementById("exposed").textContent = text;
}
</script>
<body onload="test()">
<p>CSS Grid Layout Attributes that are exposed in the CSS computed style object:</p>
<p id="exposed" style="white-space: pre"></p>
</body>
</html>
