third_party/blink/web_tests/fast/css/getComputedStyle/computed-style-grid-layout-expected.txt

CSS Grid Layout Attributes that are exposed in the CSS computed style object:

grid-auto-columns: auto;
grid-auto-flow: row;
grid-auto-rows: auto;
grid-column-end: auto;
grid-column-start: auto;
grid-row-end: auto;
grid-row-start: auto;
grid-template-areas: none;
grid-template-columns: none;
grid-template-rows: none;
